perf(dashmint-lab): defer SDK + WASM off the critical path
Lighthouse FCP/LCP were 5.1s/5.8s because the ~8MB evo-sdk WASM bundle
loaded synchronously at app boot. Switch SessionContext to a dynamic
import for createClient + IdentityKeyManager so Vite splits the SDK
into its own chunk that fetches after the shell paints. Auto-browse
is also deferred via requestAnimationFrame.

Split contract.ts: storage helpers (loadStoredContractId,
fetchContractOwnerId, ...) move to contractStorage.ts so SessionContext
can import them without dragging @dashevo/evo-sdk into the entry chunk.

Entry chunk drops from a single ~6MB-gzip bundle to 89KB gzip.
